+++ src/test/java/com/beamlytics/inventory/businesslogic/core/utils/test/avrotestobjects/TransactionsAVRO.java
@@ -38,7 +38,7 @@ public class TransactionsAVRO {
   public @org.apache.avro.reflect.Nullable long time_of_sale;
   public @org.apache.avro.reflect.Nullable int department_id;
   public @org.apache.avro.reflect.Nullable int product_id;
-  public @org.apache.avro.reflect.Nullable int product_count;
+  public @org.apache.avro.reflect.Nullable Long product_count;
   public @org.apache.avro.reflect.Nullable float price;
   public @org.apache.avro.reflect.Nullable StoreLocationAvro storeLocation;
